requirements.txt
PyYAML>=5.4.1
PySide2>=5.15.2
pyowm>=3.1.1
pytz>=2020.5
timezonefinder>=5.1.0
geopy>=2.1.0